update: MWE-23 init auto mutation
