+++ projects/server/src/controllers/stockHistoryController.js
@@ -4,131 +4,10 @@ const warehouse = require("../models/warehouse");
 const moment = require("moment");
 
 const stockHistory = {
-	getHistory: async (req, res) => {
-		try {
-			const { sort, search, warehouse_id, reference, time } = req.query;
-			const limit = 12;
-
-			const page = req?.query?.page || 1;
-			let offset = (parseInt(page) - 1) * limit;
-
-			const sortOptions = {
-				productAsc: [
-					[{ model: db.stocks }, { model: db.products }, "product_name", "ASC"],
-				],
-				productDesc: [
-					[
-						{ model: db.stocks },
-						{ model: db.products },
-						"product_name",
-						"DESC",
-					],
-				],
-				warehouseAsc: [
-					[
-						{ model: db.stocks },
-						{ model: db.warehouses },
-						"warehouse_name",
-						"ASC",
-					],
-				],
-				warehouseDesc: [
-					[
-						{ model: db.stocks },
-						{ model: db.warehouses },
-						"warehouse_name",
-						"DESC",
-					],
-				],
-				stockAfterAsc: [["stock_after", "ASC"]],
-				stockAfterDesc: [["stock_after", "DESC"]],
-				statusAsc: [["status", "ASC"]],
-				statusDesc: [["status", "DESC"]],
-				referenceAsc: [["reference", "ASC"]],
-				referenceDesc: [["reference", "DESC"]],
-				dateAsc: [["createdAt", "ASC"]],
-				dateDesc: [["createdAt", "DESC"]],
-			};
-			const sortOrder = sortOptions[sort] || sortOptions.dateDesc;
-
-			let whereClause = {};
-
-			if (search) {
-				whereClause["$stock.product.product_name$"] = {
-					[Op.like]: `%${search || ""}%`,
-				};
-			}
-
-			if (warehouse_id) {
-				whereClause["$stock.warehouse_id$"] = {
-					[Op.like]: `%${warehouse_id}%`,
-				};
-			}
-
-			if (reference) {
-				whereClause["$reference$"] = {
-					[Op.like]: `%${reference}%`,
-				};
-			}
-
-			if (time) {
-				// Apply time filter if 'time' is selected
-				whereClause[Op.and] = [
-					{
-						createdAt: { [Op.gte]: moment(time).format() },
-					},
-					{
-						createdAt: {
-							[Op.lte]: moment(time).endOf("month").format(),
-						},
-					},
-				];
-			}
-
-			const history = await db.stock_histories.findAndCountAll({
-				where: {
-					...whereClause,
-				},
-
-				include: [
-					{
-						model: db.stocks,
-						include: [
-							{ model: db.products, include: [{ model: db.product_images }] },
-							{ model: db.warehouses },
-						],
-					},
-				],
-				distinct: true,
-				order: sortOrder,
-			});
-
-			res.status(200).send({
-				count: history.count,
-				rows: history.rows.slice(offset, limit * page),
-			});
-		} catch (err) {
-			res.status(500).send({
-				message: err.message,
-			});
-		}
-	},
-
 	addStockHistory: async (dataStock, status, reference, qty) => {
 		const t = await db.sequelize.transaction();
 
 		try {
-			console.log("masuk");
-			console.log(dataStock);
-			console.log(qty);
-			console.log({
-				qty: qty - dataStock.qty,
-				status,
-				reference,
-				stock_id: dataStock.id,
-				stock_before: dataStock.qty,
-				stock_after: qty,
-			});
 			const newHistory = await db.stock_histories.create(
 				{
 					qty: qty - dataStock.qty,
@@ -143,7 +22,6 @@ const stockHistory = {
 			await t.commit();
 			return newHistory;
 		} catch (err) {
-			console.log(err);
 			await t.rollback();
 			throw err;
 		}
