+++ projects/server/src/routers/stockHistoryRouter.js
@@ -1,9 +1,10 @@
 const express = require("express");
 const router = express.Router();
-const stockController = require("../controllers").stockHistoryController;
+const historyController = require("../controllers").stockHistoryController;
+const getHistoryController = require("../controllers").getHistoryController;
 const checkRole = require("../middlewares/roleDecoder");
 
-router.get("/", checkRole.checkWAdmin, stockController.getHistory);
-router.post("/", checkRole.checkWAdmin, stockController.addStockHistory);
+router.get("/", checkRole.checkWAdmin, getHistoryController.getHistory);
+router.post("/", checkRole.checkWAdmin, historyController.addStockHistory);
 
 module.exports = router;
